+++ src/components/TodoContainer.jsx
@@ -2,6 +2,7 @@ import React from "react";
 import TodoForm from "./TodoForm";
 import TodoList from "./TodoList";
 import TodoFilter from "./TodoFilter";
+import TodoStats from "./TodoStats";
 import useTodos from "../hooks/useTodos";
 
 const TodoContainer = () => {
@@ -14,10 +15,6 @@ const TodoContainer = () => {
     filter,
     FILTERS,
   } = useTodos();
-
-  const totalTodos = todos.length;
-  const completedTodos = todos.filter((todo) => todo.completed).length;
-  const activeTodos = totalTodos - completedTodos;
 
   return (
     <div>
@@ -32,10 +29,7 @@ const TodoContainer = () => {
         toggleComplete={handleToggleComplete}
         deleteTodo={handleDeleteTodo}
       />
-      <div>
-        총 할 일 개수: {totalTodos} | 완료된 할 일: {completedTodos} | 진행 중인
-        할 일: {activeTodos}
-      </div>
+      <TodoStats todos={todos} />
     </div>
   );
 };
